[MorseSmaleQuadrangulation] Template over triangulation
