+++ core/vtk/ttkMorseSmaleQuadrangulation/ttkMorseSmaleQuadrangulation.cpp
@@ -86,8 +86,24 @@ int ttkMorseSmaleQuadrangulation::RequestData(
     ttkUtils::GetVoidPointer(sepdim), ttkUtils::GetVoidPointer(sepmask),
     ttkUtils::GetVoidPointer(seprsPoints));
 
+#define MSQUAD_EXPLICIT_CALLS(TRIANGL_CASE, TRIANGL_TYPE)                   \
+  case TRIANGL_CASE: {                                                      \
+    const auto tri = static_cast<TRIANGL_TYPE *>(triangulation->getData()); \
+    if(tri != nullptr) {                                                    \
+      res = this->execute<TRIANGL_TYPE>(*tri);                              \
+    }                                                                       \
+    break;                                                                  \
+  }
+
   int res{-1};
-  res = this->execute();
+  switch(triangulation->getType()) {
+    MSQUAD_EXPLICIT_CALLS(
+      ttk::Triangulation::Type::EXPLICIT, ttk::ExplicitTriangulation);
+    MSQUAD_EXPLICIT_CALLS(
+      ttk::Triangulation::Type::IMPLICIT, ttk::ImplicitTriangulation);
+    MSQUAD_EXPLICIT_CALLS(
+      ttk::Triangulation::Type::PERIODIC, ttk::PeriodicImplicitTriangulation);
+  }
 
   if(res != 0) {
     this->printWrn("Consider another (eigen) function, persistence threshold "
